made the code more robust and removed the animation component from the main file to better fix Skinned Mesh rendering / loading

+++ assets/Shaders/fragment.glsl
@@ -1,19 +1,18 @@
 #version 300 es
-precision mediump float;
+precision highp float;    // must match vertex shader’s float precision :contentReference[oaicite:3]{index=3}
 
-// Mirror the declaration so the linker keeps it
-uniform mat4 uBones[100];
+uniform mat4 uBones[100]; // highp default
 
-flat in vec3 vNormal;
-out vec4 FragColor;
+flat in  vec3 vNormal;    // flat qualifier to match vertex output :contentReference[oaicite:4]{index=4}
+out      vec4 FragColor;
 
 void main() {
-    // dummy read so fragment stage references it too
+    // dummy read to keep uBones alive
     if (uBones[0][0][0] > 1.0e10) {
         discard;
     }
 
-    // debug‐normal coloring
+    // debug‐normal coloring (unchanged)
     vec3 c = normalize(vNormal) * 0.5 + 0.5;
     FragColor = vec4(c, 1.0);
 }
